Move signing tests to e2e
